Fixed a bunch of warnings in safe ways

+++ libretroshare/src/file_sharing/hash_cache.cc
@@ -270,7 +270,7 @@ void HashStorage::clean()
 #endif
 
     for(std::map<std::string,HashStorageInfo>::iterator it(mFiles.begin());it!=mFiles.end();)
-        if(it->second.time_stamp + duration < (uint64_t)now)
+		if((uint64_t)(it->second.time_stamp + duration) < (uint64_t)now)
         {
 #ifdef HASHSTORAGE_DEBUG
             std::cerr << "  Entry too old: " << it->first << ", ts=" << it->second.time_stamp << std::endl ;
